build_rdev_site() for release 1.1.3
