+++ docs/404.html
@@ -1,17 +1,16 @@
 <!DOCTYPE html>
-<!-- Generated by pkgdown: do not edit by hand --><html lang="en">
+<!-- Generated by pkgdown: do not edit by hand --><html lang="en-US">
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
 <meta charset="utf-8">
 <meta http-equiv="X-UA-Compatible" content="IE=edge">
 <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 <title>Page not found (404) • quantrr</title>
-<!-- favicons --><link rel="icon" type="image/png" sizes="16x16" href="https://jabenninghoff.github.io/quantrr/favicon-16x16.png">
-<link rel="icon" type="image/png" sizes="32x32" href="https://jabenninghoff.github.io/quantrr/favicon-32x32.png">
-<link rel="apple-touch-icon" type="image/png" sizes="180x180" href="https://jabenninghoff.github.io/quantrr/apple-touch-icon.png">
-<link rel="apple-touch-icon" type="image/png" sizes="120x120" href="https://jabenninghoff.github.io/quantrr/apple-touch-icon-120x120.png">
-<link rel="apple-touch-icon" type="image/png" sizes="76x76" href="https://jabenninghoff.github.io/quantrr/apple-touch-icon-76x76.png">
-<link rel="apple-touch-icon" type="image/png" sizes="60x60" href="https://jabenninghoff.github.io/quantrr/apple-touch-icon-60x60.png">
+<!-- favicons --><link rel="icon" type="image/png" sizes="96x96" href="https://jabenninghoff.github.io/quantrr/favicon-96x96.png">
+<link rel="icon" type="”image/svg+xml”" href="https://jabenninghoff.github.io/quantrr/favicon.svg">
+<link rel="apple-touch-icon" sizes="180x180" href="https://jabenninghoff.github.io/quantrr/apple-touch-icon.png">
+<link rel="icon" sizes="any" href="https://jabenninghoff.github.io/quantrr/favicon.ico">
+<link rel="manifest" href="https://jabenninghoff.github.io/quantrr/site.webmanifest">
 <script src="https://jabenninghoff.github.io/quantrr/lightswitch.js"></script><script src="https://jabenninghoff.github.io/quantrr/deps/jquery-3.6.0/jquery-3.6.0.min.js"></script><meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
 <link href="https://jabenninghoff.github.io/quantrr/deps/bootstrap-5.3.1/bootstrap.min.css" rel="stylesheet">
 <script src="https://jabenninghoff.github.io/quantrr/deps/bootstrap-5.3.1/bootstrap.bundle.min.js"></script><link href="https://jabenninghoff.github.io/quantrr/deps/font-awesome-6.5.2/css/all.min.css" rel="stylesheet">
@@ -28,7 +27,7 @@
 
     <a class="navbar-brand me-2" href="https://jabenninghoff.github.io/quantrr/index.html">quantrr</a>
 
-    <small class="nav-text text-muted me-auto" data-bs-toggle="tooltip" data-bs-placement="bottom" title="">1.1.2</small>
+    <small class="nav-text text-muted me-auto" data-bs-toggle="tooltip" data-bs-placement="bottom" title="">1.1.3</small>
 
 
     <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbar" aria-controls="navbar" aria-expanded="false" aria-label="Toggle navigation">
@@ -77,7 +76,7 @@ Content not found. Please use links in the navbar.
 </div>
 
 <div class="pkgdown-footer-right">
-  <p>Site built with <a href="https://pkgdown.r-lib.org/" class="external-link">pkgdown</a> 2.1.1.</p>
+  <p>Site built with <a href="https://pkgdown.r-lib.org/" class="external-link">pkgdown</a> 2.1.2.</p>
 </div>
 
     </footer>
