+++ docs/pkgdown.yml
@@ -1,9 +1,9 @@
-pandoc: 3.6.2
-pkgdown: 2.1.1
+pandoc: 3.6.4
+pkgdown: 2.1.2
 pkgdown_sha: ~
 articles:
   quantrr: quantrr.html
-last_built: 2025-01-30T19:11Z
+last_built: 2025-05-14T20:20Z
 urls:
   reference: https://jabenninghoff.github.io/quantrr/reference
   article: https://jabenninghoff.github.io/quantrr/articles
